+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 (feature/polar-area-chart)
 #### Added
 - components: polar area chart
+- display version number on footer
 #### Changed
 - components: pie chart: use path instead of stroke-dasharray. There was a small white line showing on Chrome desktop with the stroke approach.