+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+(feature/polar-area-chart)
+#### Added
+- components: polar area chart
+- display version number on footer
+#### Changed
+- user: use PolarAreaChart instead of RepoCard
+- components: pie chart: use path instead of stroke-dasharray. There was a small white line showing on Chrome desktop with the stroke approach.
+- constants: update chart colors
+
 ## [0.5.0] - 2021-04-13
 #### Added
 - user: calendar line chart